Versão Final Blog_Pessoal | 10.09.2021

+++ Blog_Pessoal/src/main/java/com/example/blogpessoal/controller/TemaController.java
@@ -1,6 +1,7 @@
 package com.example.blogpessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.example.blogpessoal.model.Tema;
 import com.example.blogpessoal.repository.TemaRepository;
@@ -52,13 +54,27 @@ public class TemaController {
 
 	@PutMapping
 	public ResponseEntity<Tema> putTema(@RequestBody Tema tema) {
-		return ResponseEntity.status(HttpStatus.OK).body(temaRepository.save(tema));
+        Optional<Tema> temaUpdate = temaRepository.findById(tema.getId());
+		
+		if (temaUpdate.isPresent()) {
+			return ResponseEntity.status(HttpStatus.OK).body(temaRepository.save(tema));
+		}else{
+			throw new ResponseStatusException(
+		          	HttpStatus.NOT_FOUND, "Tema não encontrado!", null);
+		}
 
 	}
 
 	@DeleteMapping("/{id}")
 	public void deleteTema(@PathVariable long id) {
-		temaRepository.deleteById(id);
+       Optional<Tema> tema = temaRepository.findById(id);
+		
+		if (tema.isPresent()) {
+			temaRepository.deleteById(id);
+		}else{
+			throw new ResponseStatusException(
+		          	HttpStatus.NOT_FOUND, "Tema não encontrado!", null);
+		}
 	}
 	
 }
